# Use 'iolist' when building packet.