This change makes using iolist instead of contiguous binary when building packet being compressed and/or sent.

P.S. Unit tests & dialyzer passed successfully but I'm failed to wait for finish integration tests, they just hanging without any log messages. Anyway, I making this PR, in case someone find time to run all integration tests.