@joshuawscott 2017-09-19
Thanks for the PR!

Travis ran the integration tests successfully, so all looks good there. I'm unsure of the reasoning behind the change, though, could you give a rundown of the advantages of doing it this way instead of with binaries?

Cheers!