@bjhaid 2017-09-19
I think there are claims that iolist performs better when constructing data, but I believe the claims are mostly true when constructing data from a combination of charlists and binary. I'll like to see benchmarks that demonstrates that this is an improvement over what we currently have. And if my guess is wrong a rationale around why we should use iolist in lieu of the current implementation.