@vjache 2017-09-19
The matter is not that iolists claimed faster or slower. The key point here is that KafkaEx is just a kind of transport for message transmitting and those messages often produced by some serialization libraries and those libraries when construct a binary representation return iolist as a result. So, when I want to pass that result to the KafkaEx I must to convert it to a flat contiguous binary. This causes the additional usage of memory (in worst case duplicating), and guess CPU. Also if we use snappy:compress internally it is also accepts an iolist and returns flat binary, so the former binary is not need to produce.

P.S. If I'll have a time I may be perform some benchmarks and share with you.