@joshuawscott 2017-09-19
I used the following benchmark:
```
$ cat bench/create_request_bench.exs
defmodule CreateRequestBench do
  use Benchfella

  bench "create request" do
    KafkaEx.Protocol.Produce.create_request(1, "foo", %KafkaEx.Protocol.Produce.Request{
      topic: "food", partition: 0, required_acks: 1, timeout: 10, compression: :none, messages: [
        %KafkaEx.Protocol.Produce.Message{key: "", value: "hey"}
      ]
    })
  end
end
```

Current master: 
```
$ mix bench
Settings:
  duration:      1.0 s

## CreateRequestBench
[07:06:00] 1/1: create request

Finished in 1.77 seconds

## CreateRequestBench
benchmark name  iterations   average time
create request     1000000   1.58 µs/op
```

This branch:
```
$ mix bench
Settings:
  duration:      1.0 s

## CreateRequestBench
[07:02:13] 1/1: create request

Finished in 8.4 seconds

## CreateRequestBench
benchmark name  iterations   average time
create request    10000000   0.75 µs/op
```