@joshuawscott 2017-09-19
I should add that with a 64000 byte message the difference is even greater:
master: 22.71µs
use_iolists: 7.01µs