@dantswain 2017-09-19
It's tough to benchmark this.  The speed that really matters is from the function call to the data being sent out over the wire, which includes message generation and handling by the socket io.  I believe it's generally a bit more efficient to use iolists internally, we just have to be careful about making claims as to _how much_ faster.

@vjache I think this is a nice optimization.  I agree with @joshuawscott 's review, and please do try to appease ebert as much as possible (e.g., line lengths).  I know the code isn't 100% clean already but I'm trying to chip away at it.  Thanks!