@bjhaid 2017-09-19
Thanks @joshuawscott I am convinced :)

@vjache thanks for sending in the PR