@bjhaid 2017-09-19
The type of `topic`, here:

https://github.com/kafkaex/kafka_ex/blob/b0fef3290e1ece83bb7b997d6217557065a25896/lib/kafka_ex/protocol/produce.ex#L16

and here:

https://github.com/kafkaex/kafka_ex/blob/b0fef3290e1ece83bb7b997d6217557065a25896/lib/kafka_ex.ex#L297

needs to change to `iolist`

We need to make this change globally for sending requests and not just in the `produce` path.