@dantswain 2017-09-27
@vjache Do you have any time to address the PR feedback?  If not, do you mind if one of us cleans this up before we merge?