@joshuawscott 2018-05-19
Going to merge this into another branch and then fix the one test that isn't working and make a new PR.